report_results.csh: add html anchors to hashes and machines
The hashes and machine names inserted by the script in the different pages
of the Test-Results wiki are formatted as bold (`** ${hash} **:`).

Change this to level 4 HTML titles (<h4>) so that an HTML anchor is
automatically created, which makes it easier to link directly to specific
hash from a pull request.