scripts: implement '-nomodules' logic for env files
Some env files wrap the calls for loading the build and run environment
in a conditional based on the argument `-nomodules`, so that
scripts that source the env file just to get the variables defined in it
do not run slower because they have to load the environment.

This logic was copied from CICE but was not implemented in the Icepack
scripts icepack.setup and setup_run_dirs.csh.

Add the argument '-nomodules' so that the logic becomes effective.