icepack_atmo: add support for staggered atmospheric levels
In some applications, we might want the atmospheric input variables
(wind, temperature and humidity) to be given at different vertical
levels. For example, we might want to receive the winds at some level
and the scalar variables (temperature and humidity) at another.

To support this use case, add an optional argument 'zlvs' to subroutine
'icepack_atm_boundary', and pass it down to 'atmo_boundary_layer' also
as an optional argument, so that existing applications do not have to be
changed if they do not want to use this feature.

The name 'zlvs' is chosen to mirror the existing variable 'zlvl', but
with the suffix 's' for 'scalars', mimicking the existing convention for
several variables in 'atmo_boundary_layer'.

In 'atmo_boundary_layer', add new local variables 'alzs' and 'hols' to
hold quantities computed at zlvs.  Also, for clarity, rename 'alz' to
'alzm' and 'hol' to 'holm', where the 'm' suffix stands for 'momentum',
again mimicking existing convention in the subroutine.

If 'zlvs' is not passed to 'atmo_boundary_layer', it is assumed that all
variables are given at the same level (zlvl), so in that case there is
no change in the results.