icedrv_RunMod: add support for disabling thermodynamics
As noted in the previous commit, 'ktherm = -1' is documented as
disabling the thermodynamics in the Icepack documentation since 2b27a78
(Update documentation including namelist tables (#322), 2020-06-05),
although this capability does not exist in Icepack standalone.

Update icedrv_RunMod::ice_step to call some subroutines only if the
thermodynamics is enabled ('ktherm >= 0'). Namely, let prep_radiation,
biogeochemistry, step_therm2 and step_radiation only be called when
thermodynamics is enabled, while letting the rest of the subroutines
always be called. As noted in the previous commit, 'step_therm1' in
particular has to be always called since the computation of the
atmospheric stresses are done in that subroutine.