icepack_therm_shared: correct 'Tair' units in declaration comment
In subroutine icepack_therm_shared::icepack_init_trcr, 'Tair' is
documented as having units 'C' (degrees Celsius), but this is incorrect,
this variable is in Kelvins. This can be seen a few lines below where we
compute Tsfc:

    Tsfc = min(Tsmelt, Tair - Tffresh) ! deg C

where 'Tffresh' is the freezing point of fresh ice in Kelvins, 273.15,
as defined in icepack_parameters.

Fix the units in the declaration comment.

Reported-by: Frederic Dupont <frederic.dupont@canada.ca>