driver: icedrv_history: fix start time for NetCDF history output (#426)
* Trial fix for netCDF start time problem

* driver: icedrv_history: fix start time for NetCDF history output

The start time for netCDF output was hardcoded to be at 00:00:00
and there was a bug in icedrv_calendar such that idate0 was fixed
at Jan. 1. Both bugs are fixed. NOTE: if dt and istep0 are set
such that the start time contains fractional seconds, they will
be rounded down in the NetCDF output to the nearest integer.

* driver: icedrv_history: fix start time for NetCDF History output

Start of netCDF time axis is fixed at Jan. 01 00:00:00 of
init_year and time value calculation is now fixed.