doc: mention code variables relating to 'calc_dragio'
At the end of the "Ocean" section, we mention that 'dragio' can be
computed from 'iceruf_ocn' and 'thickness_ocn_layer1', but do not
mention these code variables, nor the 'calc_dragio' setting used to
activate that computation. Mention the code variables next to their
mathematical symbol, for more clarity.