Update 5-band dEdd to support test data (#472)
* Update 5-band dEdd to support test data

Add an interpolation method in icepack_shortwave.F90 for rsnw

Refactor portions of icepack_shortwave to improve robustness

Add snicar and snicartest test cases

* Update the 5-band snicar shortwave rsnw_snicar_tab interpolation

This changes answers but QC tests pass.  The new implementation checks
whether the rsnw_snicar_tab is an array of integer values with deltas
of value 1 (and sets the rsnw_snicar_chk variable).  If so, it uses a
shortcut to find the rsnw_snicar_tab bounds for interpolation, otherwise
it calls the shortwave_search routine.  In testing, the shortcut did not
change performance much, but that could change in the future.

* Clean up debug output

* Refactor the logic to control rsnw table interpolation in the shortwave.
Set a character string, rsnw_datatype, at initialization.