Update pull request template, Update conda_macos debug flag (#484)
Update pull request template to ask for additional information about the PR. This information can and will be used during the squash merge process to produce more useful commit logs.

Update Macros.conda_macos to remove ffpe-trap=invalid from debug flags.
This flag causes nf90_create to fail with the latest version of
the conda environment. This happens when debug flags are on
and a test with netcdf is turned on. I believe this is a compiler error.