Update Derecho Port (#501)
Update inteloneapi, cray, intel, nvhpc modules/compiler. Changes answers for inteloneapi and cray.

Switch default Derecho queue from main to develop to support quicker turn around and lower costs. Develop is a shared queue.

Error in the snow physics with the cray compiler needs further assessment, but suggest we merge this now as is. That error does not present itself in CICE with snow physics with the same cray compiler.