Update Derecho Cray, rename subroutines, minor cleanup
- Update Derecho Cray, add -Rp to -O2 standard optimization to address
SAME_TBS compiler bug that is triggered in limited cases.  Reported
issue to NCAR.  -O2 -Rp changes answers relative to -O2.

- Rename two public Icepack subroutines, but maintain backward compatibility,
      use icepack_therm_shared  , only: icepack_init_thermo => icepack_init_salinity
      use icepack_therm_shared  , only: icepack_init_trcr => icepack_init_enthalpy
Update the calls in Icepack driver, but also confirmed it works fine with the old
names as well as in CICE with the old names.
Closes #255

- Add a check and abort for negative values in the sqrt in computation of Tin in
function calculate_Tin_from_qin.
Closes #482

- Refactor calls to icepack_aggregate to make them consistent.  This was part of the
testing for the Derecho Cray bug, and decided to keep the implementation.

- Update comments associated with floeshape constant attribution.  Change from
Steele to Rothrock 1984.
Closes #479

- Clean up some of the variable declarations in subroutine set_state_var and module
icedrv_state, merge multiple lines to one line and shift to assumed shape arrays
where appropriate.

- Clean up implementation error in icedrv_restart.F90, subroutine restartfile.  This subroutine
was using a parameter, ntrcr, directly from icepack_tracers.  Switched that to a call to
icepack_query_tracer_sizes.