Update derecho gnu, add code coverage capability via lcov (#510)
Update derecho gnu to support code coverage testing

    Update to ncarenv/24.12 and gcc/12.4.0
    Add coverage compiler flags
    Add parse_lcov.sh script to support manually parsing lcov info files (not used in production)
    Update lcov scripting as needed

Change derecho gnu non-debug tests to -O2. This should have always been the case, but was not. This results in answers changes relative to main but only for derecho gnu with debug off.

As part of the CICE port to the new coverage tools, a bug was found in the tools that requires debug flags to be disabled when compiling with coverage flags. This was not the case when coverage was running on cheyenne. See linux-test-project/lcov#385