Use proleptic gregorian in icepack (#525)
This sets the calendar attribute correctly in history output. This update's any notes or other calendar references to support "proleptic_gregorian" as well as "gregorian" for calendar type.  Use of the calendar_type variable was also removed where it wasn't needed in the driver.  The correct cf-name for the calendar used is "proleptic_gregorian" but backward compatibility is maintained with the calendar name "gregorian". This is a companion PR to Setting calendar attribute in output to proleptic_gregorian CICE#1024.