Fix nag compiler error about kind=8 declaration, change to kind=dbl_kind (#532)
The nag error was introduced when the MOSAiC forcing was added to Icepack. It's in the Icepack driver so did not impact CICE testing. Some new variables were declared (kind=8) which isn't part of the Fortran standard even though most compilers allow it, but the nag compiler generated an error. The (kind=8) declarations were changed to (kind=dbl_kind). This should have been addressed sooner, but it should not impact solutions or results for compilers that allow it.